requirements.txt
pandas>=2.0.0
openpyxl>=3.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
python-docx>=0.8.11
